+++ nmanga/cli/manual_split.py
@@ -142,14 +142,13 @@ def _collect_archive_to_chapters(
                 continue
 
             if chapter_data not in collected_chapters:
-                if check_cbz_exist(target_path, utils.secure_filename(chapter_data)):
+                target_archive = utils.unsecure_filename(utils.secure_filename(chapter_data))
+                if check_cbz_exist(target_path, target_archive):
                     console.warning(f"[?] Skipping chapter: {chapter_data}")
                     skipped_chapters.append(chapter_data)
                     continue
                 console.info(f"[+] Creating chapter: {chapter_data}")
-                collected_chapters[chapter_data] = exporter.CBZMangaExporter(
-                    utils.secure_filename(chapter_data), target_path
-                )
+                collected_chapters[chapter_data] = exporter.CBZMangaExporter(target_archive, target_path)
 
             collected_chapters[chapter_data].add_image(path.basename(filename), archive.read(image))